requirements.txt

Django==6.0.2
django-filter==25.2
djangorestframework==3.16.1
psycopg==3.2.11
gunicorn==23.0.0
python-dotenv==1.1.1
django-cors-headers==4.9.0
Unidecode==1.4.0
Pillow==12.0.0
pytest==8.4.2
whitenoise==6.11.0
pytest-django==4.11.1
django-storages==1.14.6
boto3==1.40.56
yeeko_xlsx_export @ git+https://github.com/yeeko-org/xlsx_export@main
